add_captions {wbgcharts}	R Documentation

Add captions (title, notes, etc) to a plot or grid graphics object (grob)

Description

add_captions adds captions to a plot outside of ggplot. This is useful since the default ggplot title, subtitle and caption align with the plot area, rather than the entire figure (including axes), unlike many publishing standards.

Usage

add_captions(plot, theme, title = NULL, subtitle = NULL, note = NULL,
  source = NULL, source_url = NULL, show.logo = TRUE,
  padding = margin(0, 0, 0, 0, "pt"))

Arguments

plot
the ggplot plot object or a grid grob

theme
the ggplot theme object to use for styling captions

title
the plot title (displayed at top)

subtitle
the plot subtitle

note
the plot note

source
the plot source

source_url
the plot source URL or other location info

show.logo
if TRUE, include a logo

Details

The title and subtitle arguments correspond directly to the same in title, and appear above the plot. The note and source arguments share the work of ggplot's caption and appear below the plot. The source_url and logo appear at the very bottom of the plot, in slightly smaller text, and would usually only be shown in standalone, branded graphics (for e.g. distribution on social media) and not, for instance, within a report.

The relevant theme elements within theme are the same as would be used if the the ggplot title, subtitle and caption were used.

Value

a new gtable object which can be printed to a graphic device

Examples

library(ggplot2)
library(grid)

# Simple example for adapting a ggplot labs() call...
p <- ggplot(mtcars, aes(cyl, mpg)) + geom_point()

p + labs(
  title = "This is a title",
  subtitle = "This is a subtitle",
  caption = "This is a caption"
)

# ...to use add_captions() instead
g <- add_captions(
  plot = p,
  theme = p$theme,
  title = "This is a title",
  subtitle = "This is a subtitle",
  note = "Note: This is a note",
  source = "Note: This is a source",
  source_url = "http://where.this.is.found.com",
  show.logo = TRUE
)
grid.newpage()
grid.draw(g)

# Some relevant text attributes of ggplot themes are obeyed
p <- ggplot(mtcars, aes(cyl, mpg)) + geom_point() +
  theme_classic() + theme(plot.title = element_text(color = "blue"))

p + labs(
  title = "This is a title",
  subtitle = "This is a subtitle",
  caption = "This is a caption"
)

g <- add_captions(
  plot = p,
  theme = p$theme,
  title = "This is a title",
  subtitle = "This is a subtitle",
  note = "Note: This is a note",
  source = "Note: This is a source",
  show.logo = FALSE
)
grid.newpage()
grid.draw(g)

[Package wbgcharts version 0.1.0 Index]