bracketed_breaks {wbgcharts}	R Documentation

A breaks function for ggplot2 scales that shows the data limits

Description

This builds on the standard ggplot2 continuous breaks (extended breaks) function to make sure the data limits, given as a parameter, are included in the breaks.

Usage

bracketed_breaks(limits, at = 1, closeness_threshold = 0.1,
  breaker = scales::extended_breaks())

Arguments

limits
A pair of limits, overrides the default

at
To limit the breaks to be on 5s, 10s, etc (more of a hint)

closeness_threshold
If one of the data limit breaks and one of the default breaks will be closer than closeness_threshold * the ggplot2 calculated limits, the the nearby default break will be dropped.

Details

By default it will try to use the limits ggplot provides, but by default those are extended by the expand argument of scales, so it will use the integral ceiling and floor respectively of these. Usually this will work. If not you can provide limits to override.

[Package wbgcharts version 0.1.0 Index]