contrasting_colors {wbgcharts}	R Documentation

Select contrasting text colors, from a list, for a set of fill colors

Description

Given a (possibly named) vector of fill (background) colors, this function returns a (possibly named) vector of highest-contrast text (foreground) colors, from a list of potential text colors (usually two is sufficient).

Usage

contrasting_colors(fillcolors, textcolors = c("black", "white"),
  biases = c(0, 2))

Arguments

biases
added to the "true" contrast ratios to bias the selection

[Package wbgcharts version 0.1.0 Index]