repel {wbgcharts}	R Documentation

Repel the values in a vecto away from each other

Description

Moves the values in a vector so that they each have gap space between them. If force is FALSE (default), it will warn and do nothing if this can't be done within a space of range(x) + c(-gap, gap).

Usage

repel(x, gap = 1, force = FALSE, err = 0.01)

[Package wbgcharts version 0.1.0 Index]