str_wrap_lines {wbgcharts}	R Documentation

Wrap a vector of strings to a specified number of lines

Description

Like stringr::str_wrap() except that instead of specifying a width you specify a "height" (ie. a number of lines that each string can occupy). The longest string and "height" will jointly determine the width, which will then be applied to the other strings - so that strings are not unnecessarily wrapped.

Usage

str_wrap_lines(strings, lines = 2, indent = 0, exdent = 0,
  force = FALSE)

Arguments

force
Force strings to take up lines lines, even if they could fit the width with fewer lines

[Package wbgcharts version 0.1.0 Index]